Add: Camera Drag
The camera drag mechanic has been implemented using an UIPanGeestureRecognizer. The gesture recognizer calls a method called handlePan(recognizer:) everytime it starts, and this method makes advantage of a property called tranlation, which is a bidimencoinal vector that represents how much the user has dragged it's finger through the screen.

To drag the camera, the mathod makes the camera translates as much as the user ahs draged through the screen. In order to make the camera follow the user finger, the method has to reset the translation value everytime it's called. In this specific case the translation is reseted everytime the method is called, and it's state is either .changed or .ended.

The camera translation is limited by the largest SKTileMapNode frame, so it can't go outo of bounds.