Add: Arrow switch on touch
Add ArrowTileSwitch to manage this feature. It takes a UITouch and a SKTileMapNode as parameters to identify if the touched tile is an arrow, and if so, it set  the tile to the next arrow.

ArrowTileSwitch has a static method that verifies if the touched tile is an Arrow. If it is, the method identifies which arrow it is, and compares it's value to an enumerator that contains all possible directions. After that the method knows which is the next arrow, than it sets the arrow.